Notes on this sample script file
A few examples of supported formatting for various triggers and filters. Within a single cell, a "|" is a separater which will end up splitting the contents in that cell into a list. For example: "own|policyID|>5" is brought into the dapp (and can be translated within a custom plugin as) as: ["own", "policyID", ">5"] and can then be interpreted from there as condition of "payer must own more than 5 of policyID" and if this condition is met, the filters are applied. 

Reference the README file for this repository for more details